Increase number of buttons in page_actions before it makes a drop-down
See https://github.com/mautic/mautic/issues/751 , this would be one possible fix for that.  It may be possible to update the template so that the drop-down AND buttons are available in the HTML, and have the buttons appear/disappear based on the window width using CSS media queries.  That would probably be superior, I suppose I am not aware of why the buttons collapse into a drop-down.  Perhaps there are pages with a LOT of buttons, or it's for mobile device support?