Do not force lead with same IPh no cookie
This relates to https://www.mautic.org/community/index.php/891-using-mautic-in-mobile-app/p1#p3978

The code previously took the first lead that came from the same IP if no cookie was present. We removed that and it works well now. 

Truth is, in the proliferation of devices and people sharing public IP (e.g. office or school) we cannot make the assumption that the lead is the same. But this 'bug' is hard to catch as you needed Mautic to be used by at least two users with the same IP (e.g. office). Our application is school-based, so we identified this issue as, by design, we have same IP for potentially hundreds of users.