Bugfix: workaround for newlycreated lead flag
This workaround allows detection of a lead having recently associated
an email address in FormEvents::FORM_ON_SUBMIT listeners.

Prior to this change, a lead is created with no email upon a page hit,
and the subsequent form submission only updates it with the email; so
the "newly created" flag is not set on the form submission.