Move form chart submissions data to own method
In anticipation of additional data sources to be added to the Form
Submissions chart, the existing data set is now retrieved and added to
the chart in its own method. The getSubmissionsLineChartData method
body will thus remain tidy.

In the newly defined 'setChartDataSubmissionsCount' method, the result
limiting query modification has been replaced with a call to the
recently added 'limitQueryToCreator', as will be done in subsequently
added chart data methods.