Alias DoNotContact entity as DNC in LeadModel
The model counterpart to the DoNotContact Entity is also named
DoNotContact, breaking with the convention of Model classes being
suffixed as such. This can cause a classname collision within the
LeadModel class which imports the DoNotContact entity class. To
prevent collision, the import is aliased as "DNC" and all mentions of
the DoNotContact classname within LeadModel have been replaced with
the alias.