Support graceful timeout option on ConsumeQueueCommand
Added an option (--timeout=TIMEOUT, -t TIMEOUT) takes a number of
seconds to be used as the graceful execution time on the Queue
Consumer.

The value is then passed as an argument to the Queue Service's
consumeFromQueue method, where it is passed as a constructor argument
for the QueueEvent to be dispatched. Then, the RabbitMqSubscriber uses
the event's timeout property to maybe call the consumer's
setGracefulMaxExecutionDateTimeFromSecondsInTheFuture method before
calling $consumer->consume(...).

This will limit the execution time of the ConsumeQueueCommand by
allowing it to exit gracefully after TIMEOUT seconds have elapsed.