Define SendGridMail Categories part
Like the SendGridMailMetadata, SendGridMailPersonalization,
etc. classes, the SendGridMailCategories handles management of an
attribute of a SendGrid\Mail instance through the SendGridApiMessage
class.

Unlike the other message part classes, this one uses the
event_dispatcher to allow interaction from plugins with subscribers
listening for the SendGridMailEvents::ADD_CATEGORIES event.

Skips dispatching events and adding categories on messages that are
not instances of MauticMessage.