Possible fix for #7294
When the configuration is updated a `ConfigEvent` is dispatched carrying both the original form submission data and a normalized version. In the normalized version files are represented by `UploadedFile` instances.

The event is being picked up by the audit logger which attempts to store it in the database by serializing the configuration into a string. According to https://github.com/symfony/symfony-docs/pull/8180 `UploadedFile` instances cannot be serialized and throw an exception, which possibly what is happening in #7294.

With the above PR I am passing the normalized data through a filter which replaces all `UploadedFile` instances (recursively) with their contents (maybe the content is not the best idea?) when being passed to the event, allowing for the whole thing to be further serialized.