Add query to apply MySQL defaults.
sql_mode does not include "ONLY_FULL_GROUP_BY" by default in stock MySQL, but it causes various problems in Mautic if enabled as Mautic is not compatible with that strict convention. innodb_default_row_format is "dynamic" by default in MySQL 5.7+ but there is no downside to setting this early with ~5.6 as it allows a smoother upgrade path, performance benefits, and reduced index limitations. We could set these (at the session level) during installation/migration/plugin-update automatically in the future and remove the need to explain this. Re: #5328 #6471