Add support for dns-based redis host discovery
PRedis only uses one redis host even if it is provided
with a hostname resolving to multiple addresses. To solve that
we have to resolve the hostname ourselves and provide an array
of available hosts to PRedis. This is usefull in HA setups where
we want to continue working even if one of the hosts fails.