Creating new model for email stat entity save
if we manage to send all Stat entity save through this model then we will be able to allow subscribers to do custom actions on the Stat entity pre and post save.

New EmailStatEvent that will be dispatched on State pre and post save.

Using EmailStatModel in the EmailModel instead of the repository to save Stat entities

Using EmailStatModel in the StatHelper instead of the repository to save Stat entities

Using EmailStatModel in the Bounce instead of the repository to save Stat entities

Using EmailStatModel in the Reply instead of the repository to save Stat entities

Fixing SendEmailToContactTest that used live instance of StatHelper

The StatHelper's dependencies has changed. I refactored the test a bit to be more DRY and use shorter syntax to improve readablity and maintainability

Using EmailStatModel in the TransportCallback instead of the repository to save Stat entities

Service dependencies updated

The change in the MailHelper covered with a test

Use EmailModel (EmailStatModel) to save email Stat entity to trigger the new events.

Also refactored to return early + fixed probably forgotten debug where the retry if was always true

Method PHPUnit\Framework\TestCase::once() invoked with 1 parameter, 0  ​required.

Adding option to record changed fields in the Stat entity

This is useful for subscribers listening for ON_EMAIL_STAT_POST_SAVE to know what was changed.

Property comment fix

return type added

Co-authored-by: Saurabh Gupta <48244990+dadarya0@users.noreply.github.com>

return type added

Co-authored-by: Saurabh Gupta <48244990+dadarya0@users.noreply.github.com>

Using constants for retry counts

CS fix

Introducing DateTimeHelper::FORMAT_DB constant so we wouldn't have to repeat the 'Y-m-d H:i:s' string format in hundreds of places.