Allow for preconfigured cache & log directories
Currently the checkstep takes a very biased approach where logs and cache are placed. This allows for these configurations to be overridden in local parameters and be loaded in before installation time. 

Without this change, installing Mautic is not possible without first setting these folders to the "biased" location and only afterwards they can be moved.