We cannot throw an exception as...
it results in an infinite loop while finding contacts that and adding members. There are always more contacts to be added if we skip some.

So instad, we'll always create a new campaign member. If the camaign is repeatable then we set the member as removed. If it is repeatable then we created with second rotation.