Removed cascadeDetach() from Lead to Stage
Addresses #8951 and similar situations in which a copy of a Stage is generated by doctrine during API calls due to the associated Stage getting detached when a Lead is detached.
A single Lead is not the owner of a Stage so, besides being an easy solution, this seems appropriate. However, it is a global change of the configuration and could affect other places.