Tests for sending emails through Mautic's API
Add test cases for the following couple of API endpoints:

POST /emails/ID/send
Send a segment email to linked segment(s).

POST /emails/ID/contact/CONTACT_ID/send
Send a predefined email to existing contact.

Tests are made with mockup HTTP queries using the `request()` method of
a symfony's `KernelBrowser` object whose implementation is delegated to
Mautic's `AppKernel`.
As for emails that would otherwise be sent to recipients, these are caught
in a mockup SMTP transport object which doesn't actually send any emails,
but rather it stores a copy of the last email attempted which is made
accessible to the testing facilities for verification.