Fix issue #11267 - Sending emails via API should respect useOwnerAsMailer
Sending an email using the API endpoint `POST /emails/ID/contact/CONTACT_ID/send`
does not respect the email option `useOwnerAsMailer`.

The expected behaviour is for an email that would be sent by Mautic
it should appear to be sent by the `owner` of the contact who is the recipient,
i.e. for every sent email with the `useOwnerAsMailer` option enabled, the
sender's name and sender's email should belong to the `owner'.

This is a small bug where the owner is not provided in the contact
information passed to the mail sender.